Support different actors types for autonaming
