+++ posthog/api/insight.py
@@ -41,7 +41,8 @@ from posthog.hogql.timings import HogQLTimings
 from posthog import schema
 from posthog.api.documentation import extend_schema, extend_schema_field, extend_schema_serializer
 from posthog.api.forbid_destroy_model import ForbidDestroyModel
-from posthog.api.insight_suggestions import generate_insight_metadata, get_insight_analysis, get_insight_suggestions
+from posthog.api.insight_metadata import generate_insight_metadata
+from posthog.api.insight_suggestions import get_insight_analysis, get_insight_suggestions
 from posthog.api.insight_variable import map_stale_to_latest
 from posthog.api.monitoring import Feature, monitor
 from posthog.api.query_coalescer import QueryCoalescingMixin
@@ -226,6 +227,13 @@ def capture_legacy_api_call(request: request.Request, team: Team) -> None:
     except Exception as e:
         logging.exception(f"Error in capture_legacy_api_call: {e}")
         pass
+
+
+SUPPORTED_METADATA_ACTOR_SOURCES = (
+    schema.InsightActorsQuery,
+    schema.FunnelsActorsQuery,
+    schema.StickinessActorsQuery,
+)
 
 
 class QuerySchemaParser(JSONParser):
@@ -1570,9 +1578,9 @@ When set, the specified dashboard's filters and date range override will be appl
         try:
             if kind == "ActorsQuery":
                 query = schema.ActorsQuery.model_validate(query_data)
-                if not query.source or not hasattr(query.source, "source"):
+                if not isinstance(query.source, SUPPORTED_METADATA_ACTOR_SOURCES):
                     return Response(
-                        {"error": "ActorsQuery must have an insight source (e.g. InsightActorsQuery)"},
+                        {"error": "ActorsQuery must have a supported insight source (e.g. InsightActorsQuery)"},
                         status=status.HTTP_400_BAD_REQUEST,
                     )
             else:
